DRR - Cpptraj: Fix implicit conversion of int to size_t. Not technically a problem since it gets converted back to int in ClusterSieve, but better to be correct and consistent.

+++ src/DataSet_Cmatrix.cpp
@@ -11,7 +11,7 @@ void DataSet_Cmatrix::PrintElements() const {
 /** Set up sieving info as necessary and set up cluster based on actual
   * number of frames to be clustered.
   */
-int DataSet_Cmatrix::SetupWithSieve(ClusterDist* CdistIn, size_t sizeIn, size_t sieveIn, int iseed)
+int DataSet_Cmatrix::SetupWithSieve(ClusterDist* CdistIn, size_t sizeIn, int sieveIn, int iseed)
 {
   if (CdistIn == 0) {
     mprinterr("Internal Error: DataSet_Cmatrix::SetupWithSieve called with empty ClusterDist.\n");
